refactor: update component items
