+++ src/components/AuthorList/AuthorListCards/index.scss
@@ -1,0 +1,71 @@
+@import '../../../assets/styles/variables.scss';
+
+$width-full: 100%;
+
+.row {
+  padding-top: 1rem;
+  justify-content: center;
+}
+
+.card-img {
+  cursor: pointer;
+  max-width: $width-full;
+  width: fit-content;
+  transition: transform 1s cubic-bezier(0.23, 1, 0.32, 1);
+
+  &:hover {
+    transform: scale(1.07);
+    transition: transform 1s cubic-bezier(0.23, 1, 0.32, 1);
+  }
+}
+
+.mb-4 {
+  height: 650px;
+  margin-bottom: 0 !important;
+
+  .card {
+    padding-top: 15px;
+    height: 93%;
+    border: none;
+    text-align: center;
+    align-items: center;
+    color: white;
+    background: url('../../../assets/images/fon-main-block2.jpg') no-repeat;
+    background-size: cover;
+    background-attachment: fixed;
+  }
+}
+
+.card-body {
+  padding: 1rem;
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: center;
+  width: $width-full;
+  background: $color-bg-main-block2;
+}
+
+p {
+  width: $width-full;
+  margin-bottom: 0.75rem;
+}
+
+.no-result {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  height: 10rem;
+}
+
+@media (max-width: 770px) {
+  .card-title {
+    font-size: 1rem;
+  }
+
+  .btn {
+    font-size: 0.75rem;
+    padding: 0.3rem 0.55rem;
+  }
+}
